+++ src/conversion/pipelineStringToJsonSync.ts
@@ -1,17 +1,21 @@
 import { spaceTrim } from 'spacetrim';
 import type { Writable, WritableDeep } from 'type-fest';
+import { isPipelineImplementingInterface } from '../_packages/core.index';
 import type { ParameterCommand } from '../commands/PARAMETER/ParameterCommand';
 import { sectionCommandParser } from '../commands/SECTION/sectionCommandParser';
 import { getParserForCommand } from '../commands/_common/getParserForCommand';
 import { parseCommand } from '../commands/_common/parseCommand';
-import type { $PipelineJson } from '../commands/_common/types/CommandParser';
-import type { $TaskJson } from '../commands/_common/types/CommandParser';
-import type { CommandBase } from '../commands/_common/types/CommandParser';
-import type { PipelineHeadCommandParser } from '../commands/_common/types/CommandParser';
-import type { PipelineTaskCommandParser } from '../commands/_common/types/CommandParser';
+import type {
+    $PipelineJson,
+    $TaskJson,
+    CommandBase,
+    PipelineHeadCommandParser,
+    PipelineTaskCommandParser,
+} from '../commands/_common/types/CommandParser';
 import { RESERVED_PARAMETER_NAMES } from '../config';
 import { ParseError } from '../errors/ParseError';
 import { UnexpectedError } from '../errors/UnexpectedError';
+import { FORMFACTOR_DEFINITIONS } from '../formfactors';
 import type { ParameterJson } from '../pipeline/PipelineJson/ParameterJson';
 import type { PipelineJson } from '../pipeline/PipelineJson/PipelineJson';
 import type { ScriptTaskJson } from '../pipeline/PipelineJson/ScriptTaskJson';
@@ -19,9 +23,7 @@ import type { TaskJson } from '../pipeline/PipelineJson/TaskJson';
 import type { PipelineString } from '../pipeline/PipelineString';
 import type { ScriptLanguage } from '../types/ScriptLanguage';
 import { SUPPORTED_SCRIPT_LANGUAGES } from '../types/ScriptLanguage';
-import type { number_integer } from '../types/typeAliases';
-import type { number_positive } from '../types/typeAliases';
-import type { string_name } from '../types/typeAliases';
+import type { number_integer, number_positive, string_name } from '../types/typeAliases';
 import { extractAllListItemsFromMarkdown } from '../utils/markdown/extractAllListItemsFromMarkdown';
 import { extractOneBlockFromMarkdown } from '../utils/markdown/extractOneBlockFromMarkdown';
 import { flattenMarkdown } from '../utils/markdown/flattenMarkdown';
@@ -56,7 +58,8 @@ export function pipelineStringToJsonSync(pipelineString: PipelineString): Pipeli
         pipelineUrl: undefined /* <- Note: Putting here placeholder to keep `pipelineUrl` on top at final JSON */,
         bookVersion: undefined /* <- Note: By default no explicit version */,
         description: undefined /* <- Note: [🍙] Putting here placeholder to keep `description` on top at final JSON */,
-        formfactorName: 'GENERIC',
+        formfactorName:
+            undefined /* <- Note: [🍙] + When undefined at the end of the parsing, it will be set to 'GENERIC' */,
         parameters: [],
         tasks: [],
         knowledgeSources: [],
@@ -300,7 +303,36 @@ export function pipelineStringToJsonSync(pipelineString: PipelineString): Pipeli
     }
 
     // =============================================================
-    // Note: 4️⃣ Prepare unique section names with indexes when needed
+    // Note: 4️⃣ Implicit and default formfactor
+
+    for (const formfactorDefinition of FORMFACTOR_DEFINITIONS) {
+        // <- Note: [♓️][💩] This is the order of the formfactors, make some explicit priority
+
+        const { name, pipelineInterface } = formfactorDefinition;
+
+        const isCompatible = isPipelineImplementingInterface({
+            pipeline: {
+                ...$pipelineJson,
+                formfactorName: name,
+                // <- Note: `formfactorName` has no role in `isPipelineImplementingInterface`
+                //           but it is needed to satisfy the typescript
+            },
+            pipelineInterface,
+        });
+
+        if (isCompatible) {
+            $pipelineJson.formfactorName = name;
+            break;
+        }
+    }
+
+    // Note: [🔆] If formfactor is still not set, set it to 'GENERIC'
+    if ($pipelineJson.formfactorName === undefined) {
+        $pipelineJson.formfactorName = 'GENERIC';
+    }
+
+    // =============================================================
+    // Note: 5️⃣ Prepare unique section names with indexes when needed
 
     const sectionCounts: Record<
         string_name,
@@ -332,7 +364,7 @@ export function pipelineStringToJsonSync(pipelineString: PipelineString): Pipeli
     };
 
     // =============================================================
-    // Note: 5️⃣ Process each section of the pipeline
+    // Note: 6️⃣ Process each section of the pipeline
     for (const section of pipelineSections) {
         // TODO: Parse section's description (the content out of the codeblock and lists)
 
@@ -537,7 +569,7 @@ export function pipelineStringToJsonSync(pipelineString: PipelineString): Pipeli
     }
 
     // =============================================================
-    // Note: 6️⃣ Mark parameters as INPUT if not explicitly set
+    // Note: 7️⃣ Mark parameters as INPUT if not explicitly set
     if ($pipelineJson.parameters.every((parameter) => !parameter.isInput)) {
         for (const parameter of $pipelineJson.parameters) {
             const isThisParameterResulting = $pipelineJson.tasks.some(
@@ -545,22 +577,27 @@ export function pipelineStringToJsonSync(pipelineString: PipelineString): Pipeli
             );
             if (!isThisParameterResulting) {
                 parameter.isInput = true;
+                // <- TODO: [💔] Why this is making typescript error in vscode but not in cli
+                //        > Type 'true' is not assignable to type 'false'.ts(2322)
+                //        > (property) isInput: false
+                //        > The parameter is input of the pipeline The parameter is NOT input of the pipeline
             }
         }
     }
 
     // =============================================================
-    // Note: 7️⃣ Mark all non-INPUT parameters as OUTPUT if any OUTPUT is not set
+    // Note: 8️⃣ Mark all non-INPUT parameters as OUTPUT if any OUTPUT is not set
     if ($pipelineJson.parameters.every((parameter) => !parameter.isOutput)) {
         for (const parameter of $pipelineJson.parameters) {
             if (!parameter.isInput) {
                 parameter.isOutput = true;
+                // <- TODO: [💔]
             }
         }
     }
 
     // =============================================================
-    // Note: 8️⃣ Cleanup of undefined values
+    // Note: 9️⃣ Cleanup of undefined values
     $pipelineJson.tasks.forEach((tasks) => {
         for (const [key, value] of Object.entries(tasks)) {
             if (value === undefined) {
@@ -578,7 +615,11 @@ export function pipelineStringToJsonSync(pipelineString: PipelineString): Pipeli
     // =============================================================
 
     // TODO: [🍙] Maybe do reorder of `$pipelineJson` here
-    return $asDeeplyFrozenSerializableJson('pipelineJson', $pipelineJson);
+    return $asDeeplyFrozenSerializableJson('pipelineJson', {
+        ...$pipelineJson,
+        formfactorName: 'GENERIC',
+        // <- Note: [🔆] This is redundant to satisfy the typescript
+    });
 }
 
 /**